Remove nn package.
- Move ActivationFunctions to core
- Examples are already in deepwit
- Move optimziers to optimizer package
