+++ examples/src/main/scala/basic/LogisticRegression.scala
@@ -3,8 +3,8 @@ package examples.basic
 import dimwit.*
 import dimwit.Conversions.given
 import dimwit.autodiff.*
-import nn.*
-import nn.ActivationFunctions.{sigmoid, relu}
+import dimwit.optimizer.GradientDescent
+import dimwit.nn.ActivationFunctions.{sigmoid, relu}
 import dimwit.random.Random
 import dimwit.stats.Normal
