--- nn/src/main/scala/nn/package.scala
@@ -1,1 +1,0 @@
-package object nn {}
